requirements.txt
amqp==2.6.1
asgiref==3.3.1
billiard==3.6.4.0
black==21.10b0
celery==4.4.6
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
comtypes==1.1.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.1.7
django-appconf==1.0.5
django-braces==1.15.0
django-celery-email==3.0.0
django-cors-headers==3.5.0
django-environ==0.8.1
django-filter==21.1
django-oauth-toolkit==1.5.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==5.0.0
docx-mailmerge==0.5.0
drf-social-oauth2==1.0.8
future==0.18.2
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
jwcrypto==1.0
kombu==4.6.11
lxml==4.6.4
MarkupSafe==1.1.1
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
pdf2image==1.16.0
Pillow==8.1.2
platformdirs==2.4.0
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
regex==2021.11.10
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
tomli==1.2.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.3
vine==1.3.0
wrapt==1.13.3
